Updated the flow and fixed some HTML syntax
Just update the flow since the Top/Front/Side needs to be selected before you choose the file.  Also added labels to the input options so you can click the words instead of just the ratio buttons.  Added a title and updated the font to be correct on all browsers.